adds day17 part2
This is an ugly copy/paste of the part1 with some small
modifications to handle the `w` dimension.
We should probably have refactored part1 to be more generic and be
able to handle N dimensions.
But this problem was a bit boring, and didn't want to spend to much
time on it... So 🙈