initialize day20 and implement tile parser
We will store only the id and the border of a tile, we don't really
care about the inner content of tiles.
And in order to save space and have efficient comparisons when
trying to match tiles, we will transform the string composed by
dashes and dots into binaries (a dash is a 1 and a . is a 0).