solve part1
Don't need to do the entire jigsaw, while looking at the data and
playing with matches, I realized that in the example and the given
input, each time we have 4 tiles that can be match in only 4
directions. So it means they are the corners.

Also I realized while looking at the given example, that some of the
tiles were flipped and rotated 😈