adds methods to flip or rotate matrixes
Those method will mutate the given matrix. We will need those
methods when we will need to draw the tiles after finishing the
jigsaw to build each tile and concatenate them into the big drawing,
we will also need those methods to flip and rotate the big drawing
when looking for sea monster.