solve part1
Just loop over the food, and associate possible inxgredients to
each allergen. And everytime we do the intersection of possibel
ingredients. Like that at some point we might isolate matches: a
match being a couple allergen, ingredient.
Then we remove the known ingredient and the known allergen, and we
continue to reduce the food till there is no more possible
reduction.

Finally all the safe ingredient are in food where we don't have any
more allergen.